+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Add a `glance_metric` component ([PR #4452](https://github.com/alphagov/govuk_publishing_components/pull/4452))
 * Use component wrapper on org logo component ([PR #4458](https://github.com/alphagov/govuk_publishing_components/pull/4458))
 * Use component wrapper on notice component ([PR #4444](https://github.com/alphagov/govuk_publishing_components/pull/4444))
+* Add component wrapper to the metadata component ([PR #4442](https://github.com/alphagov/govuk_publishing_components/pull/4442))
 
 ## 46.0.0